Make json.h a single header library.
This brings the implementation of json.c into json.h as weak symbols, using a similar approach I've used elsewhere. It means that compile time will be longer when using json.h, but in practice everyone generally includes it in a single file.
